data-science/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.8.2
